FileTract/requirements.txt
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
opencv-python-headless>=4.8.0
scipy>=1.11.0
numpy>=1.24.0
Flask>=3.0.0
Flask-CORS>=4.0.0
gunicorn>=21.0.0
